+++ packaging/applesmc-t2touch/applesmc.c
@@ -1024,7 +1024,7 @@ static int applesmc_init_smcreg_try(struct applesmc_device *smc)
 	s->key_count = count;
 
 	if (!s->cache)
-		s->cache = kzalloc_objs(*s->cache, s->key_count);
+		s->cache = kcalloc(s->key_count, sizeof(*s->cache), GFP_KERNEL);
 	if (!s->cache)
 		return -ENOMEM;
 
@@ -1893,7 +1893,7 @@ static int applesmc_create_nodes(struct applesmc_device *smc,
 	int ret, i;
 
 	for (grp = groups; grp->format; grp++) {
-		grp->nodes = kzalloc_objs(*node, num + 1);
+		grp->nodes = kcalloc(num + 1, sizeof(*node), GFP_KERNEL);
 		if (!grp->nodes) {
 			ret = -ENOMEM;
 			goto out;